Move ref readonly parameter tests to one file
